backend/requirements.txt
fastapi
uvicorn[standard]

SQLAlchemy>=2.0
psycopg2-binary

pydantic
email-validator
python-dotenv

python-jose
passlib[bcrypt]
bcrypt<4.1

alembic
httpx
openai
neo4j
redis
qdrant-client
fastembed
python-docx
